+++ src/CxbxKrnl/EmuXTL.h
@@ -39,13 +39,13 @@
 namespace XTL
 {
     #include "core/HLE/XAPI/Xapi.h"
-    #include "EmuD3D8.h"
-    #include "EmuD3D8\Convert.h"
-    #include "EmuD3D8\VertexBuffer.h"
-    #include "EmuD3D8\PushBuffer.h"
-    #include "EmuD3D8\VertexShader.h"
-	#include "EmuD3D8\PixelShader.h"
-    #include "EmuD3D8\State.h"
+    #include "core/HLE/D3D8/Direct3D9/Direct3D9.h"
+    #include "core/HLE/D3D8/XbConvert.h"
+    #include "core/HLE/D3D8/XbVertexBuffer.h"
+    #include "core/HLE/D3D8/XbPushBuffer.h"
+    #include "core/HLE/D3D8/XbVertexShader.h"
+	#include "core/HLE/D3D8/XbPixelShader.h"
+    #include "core/HLE/D3D8/XbState.h"
     #include "core/HLE/XAPI/OHCI/XInput/DInput.h"
     #include "core/HLE/DSOUND/DirectSound/DirectSound.hpp"
     #include "core/HLE/XONLINE/XOnline.h"
