+++ src/core/HLE/D3D8/Direct3D9/Direct3D9.cpp.unused-patches
