+++ src/core/HLE/D3D8/XbPushBuffer.h
@@ -7,7 +7,7 @@
 // *  `88bo,__,o,    oP"``"Yo,  _88o,,od8P   oP"``"Yo,
 // *    "YUMMMMMP",m"       "Mm,""YUMMMP" ,m"       "Mm,
 // *
-// *   Cxbx->Win32->CxbxKrnl->EmuD3D8->State.h
+// *   core->HLE->D3D8->XbPushBuffer.h
 // *
 // *  This file is part of the Cxbx project.
 // *
@@ -26,33 +26,30 @@
 // *  If not, write to the Free Software Foundation, Inc.,
 // *  59 Temple Place - Suite 330, Bostom, MA 02111-1307, USA.
 // *
-// *  (c) 2002-2004 Aaron Robinson <caustik@caustik.com>
+// *  (c) 2002-2003 Aaron Robinson <caustik@caustik.com>
 // *
 // *  All rights reserved
 // *
 // ******************************************************************
-#ifndef STATE_H
-#define STATE_H
+#ifndef XBPUSHBUFFER_H
+#define XBPUSHBUFFER_H
 
-#define X_D3DRS_UNSUPPORTED (X_D3DRS_LAST + 1)
+extern int DxbxFVF_GetNumberOfTextureCoordinates(DWORD dwFVF, int aTextureIndex);
+extern UINT DxbxFVFToVertexSizeInBytes(DWORD dwFVF, BOOL bIncludeTextures);
 
-// XDK version independent renderstate table, containing pointers to the original locations.
-extern DWORD *EmuMappedD3DRenderState[X_D3DRS_UNSUPPORTED]; // 1 extra for the unsupported value
+extern void CxbxDrawIndexed(CxbxDrawContext &DrawContext);
+extern void CxbxDrawPrimitiveUP(CxbxDrawContext &DrawContext);
 
-struct X_Stream {
-    DWORD Stride;
-    DWORD Offset;
-    XTL::X_D3DVertexBuffer *pVertexBuffer;
-};
+extern void EmuExecutePushBuffer
+(
+    X_D3DPushBuffer       *pPushBuffer,
+    X_D3DFixup            *pFixup
+);
 
-// EmuD3DDeferredRenderState
-extern DWORD *EmuD3DDeferredRenderState;
-
-// EmuD3DDeferredTextureState
-extern DWORD *EmuD3DDeferredTextureState;
-
-extern void EmuUpdateDeferredStates();
-
-extern void CxbxUpdateNativeD3DResources();
+extern void EmuExecutePushBufferRaw
+(
+	void *pPushData,
+	uint32_t uSizeInBytes
+);
 
 #endif
